Added support for including and excluding files from the list of files discovered
